Title: 계단 오르기, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 241009/계단 오르기/climbing-stairs.cpp
@@ -5,7 +5,7 @@
 
 using namespace std;
 
-int routes[MAX_SIZE];
+long long routes[MAX_SIZE];
 int N;
 
 int stair (int floor) {
